requirements.txt
pytest~=8.3.5
selenium~=4.30.0
webdriver-manager~=4.0.2
requests~=2.32.3
allure-python-commons~=2.14.1
